Tools-and-Frameworks/MCP/CustomMCPServer/requirements.txt
langchain-mcp-adapters[fastmcp,http]
langchain-openai
langgraph
python-dotenv
uvicorn
